[Deno] Convert unit tests, use tincan
